+++ common/src/main/java/dev/gimme/gimmehardcore/domain/config/GeneralConfig.java
@@ -5,7 +5,6 @@ public abstract class GeneralConfig {
     public static GeneralConfig INSTANCE;
 
     public abstract boolean pickaxeRequirements();
-    public abstract float getNaturalRegenerationSpeedMultiplier();
     public abstract int getMaxIronGolemIngotDrops();
     public abstract boolean disableLavaBucket();
     public abstract float getEnderDragonMaxExplosionDamage();