+++ common/src/main/java/dev/gimme/gimmehardcore/domain/config/RegenConfig.java
@@ -1,0 +1,16 @@
+package dev.gimme.gimmehardcore.domain.config;
+
+public abstract class RegenConfig {
+
+    public static RegenConfig INSTANCE;
+
+    public abstract float getNaturalRegenerationSpeedMultiplier();
+
+    public abstract float getHealAmount();
+    public abstract float getSecondsBetweenHeals();
+    public abstract float getCampfireRequiredPlayerRatio();
+    public abstract float getCampfireRange();
+    public abstract boolean isCampfireOnlyAboveSeaLevel();
+    public abstract float getCampfireHealExhaustion();
+    public abstract float getCampfireMaxHealToPercentage();
+}