+++ neoforge/src/main/java/dev/gimme/gimmehardcore/config/NeoForgeGeneralConfig.java
@@ -13,13 +13,6 @@ public class NeoForgeGeneralConfig extends GeneralConfig {
                 Hint: You can still acquire blocks through explosions.""")
             .define("pickaxeRequirements", true);
 
-    private static final ModConfigSpec.DoubleValue NATURAL_REGENERATION_SPEED_MULTIPLIER = BUILDER
-            .comment("""
-                    Multiplier for natural health regeneration speed. Note: this also makes being saturated not heal you faster.
-                    For example, setting this to 0.1 makes it take 40 seconds per half heart instead of 4 seconds.
-                     Vanilla: set to -1""")
-            .defineInRange("naturalRegenerationSpeedMultiplier", 0.1, -1.0, 1.0);
-
     private static final ModConfigSpec.IntValue MAX_IRON_GOLEM_INGOT_DROPS = BUILDER
             .comment("""
                     Maximum number of iron ingots dropped by iron golems. The minimum becomes 2 less than this value.
@@ -39,11 +32,6 @@ public class NeoForgeGeneralConfig extends GeneralConfig {
     @Override
     public boolean pickaxeRequirements() {
         return PICKAXE_REQUIREMENTS.get();
-    }
-
-    @Override
-    public float getNaturalRegenerationSpeedMultiplier() {
-        return NATURAL_REGENERATION_SPEED_MULTIPLIER.get().floatValue();
     }
 
     @Override