+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added Campfire passive healing feature
+
+### Fixed
+
+- Various bug fixes and improvements
+
 ## [v1.21.11-0.1.0] - 2026-01-17
 
 ### Added